# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_httpapi.createTeamRequest:
    properties:
      name:
        type: string
    type: object
  internal_httpapi.errorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_httpapi.errorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_httpapi.errorInfo'
    type: object
...
  internal_httpapi.inviteRequest:
    properties:
      email:
        type: string
      role:
        type: string
    type: object
...
  internal_httpapi.membersResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/task_internal_model.TeamMember'
        type: array
    type: object
...
  internal_httpapi.statusResponse:
    properties:
      status:
        type: string
    type: object
...
  internal_httpapi.teamResponse:
    properties:
      team:
        $ref: '#/definitions/task_internal_model.Team'
    type: object
  internal_httpapi.teamsResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/task_internal_model.Team'
        type: array
    type: object
  internal_httpapi.updateMemberRoleRequest:
    properties:
      role:
        type: string
    type: object
...
  task_internal_model.Role:
    enum:
    - owner
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleOwner
    - RoleAdmin
    - RoleMember
...
  task_internal_model.Team:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  task_internal_model.TeamMember:
    properties:
      email:
        type: string
      joined_at:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/task_internal_model.Role'
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'REST API для управления задачами внутри команд: ролевая модель, история
    изменений, кеширование списков задач и SQL-отчёт.'
  title: Task Manager API
  version: "1.0"
paths:
...
  /teams:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.teamsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: List my teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Team payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.createTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.teamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a team
      tags:
      - teams
  /teams/{team_id}/invite:
    post:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Invite payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.inviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Invite a user to a team
      tags:
      - teams
  /teams/{team_id}/members:
    get:
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.membersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: List team members
      tags:
      - teams
  /teams/{team_id}/members/{user_id}:
    delete:
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Target user ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove a team member
      tags:
      - teams
    patch:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Target user ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: New role (admin | member)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.updateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Change a member's role
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"